refactor(urls): split pattern.rs and reverse.rs by responsibility (#4311)
Decompose pattern.rs (1732 lines) and reverse.rs (1614 lines) into
responsibility-based submodules under pattern/ and reverse/ following
Rust 2024 Edition module conventions. Public API preserved via pub use.

Fixes #4311
